.pr-reviewer/requirements.txt
# ================================================================
# PR Reviewer Agent - GitHub Actions Requirements
# Optimized for CI/CD pipeline execution
# ================================================================

# ----------------------------------------------------------------
# Core Framework
# ----------------------------------------------------------------
agentfield>=0.1.35                # AgentField SDK

# ----------------------------------------------------------------
# AI/LLM
# ----------------------------------------------------------------
anthropic>=0.21.0                # Claude API
httpx>=0.27.0                    # HTTP client

# ----------------------------------------------------------------
# GitHub Integration
# ----------------------------------------------------------------
PyGithub>=2.1.1                  # GitHub API
GitPython>=3.1.40                # Git operations

# ----------------------------------------------------------------
# Python Language Tools
# ----------------------------------------------------------------
pytest>=8.0.0                    # Testing
pytest-json-report>=1.5.0        # JSON reports
pytest-cov>=4.1.0                # Coverage
pylint>=3.0.0                    # Linting
ruff>=0.3.0                      # Fast linter
mypy>=1.8.0                      # Type checking
black>=24.0.0                    # Formatter
isort>=5.13.0                    # Import sorter
bandit>=1.7.7                    # Security scanner
radon>=6.0.1                     # Complexity

# ----------------------------------------------------------------
# Configuration & Data
# ----------------------------------------------------------------
pyyaml>=6.0.1                    # YAML parser
python-dotenv>=1.0.0             # Environment variables
jsonschema>=4.21.0               # JSON validation
pydantic>=2.6.0                  # Data validation

# ----------------------------------------------------------------
# CLI & Output
# ----------------------------------------------------------------
rich>=13.7.0                     # Terminal output
click>=8.1.7                     # CLI framework

# ----------------------------------------------------------------
# Async & HTTP
# ----------------------------------------------------------------
aiofiles>=23.2.1                 # Async file I/O
aiohttp>=3.9.0                   # Async HTTP
requests>=2.31.0                 # HTTP library
nest-asyncio>=1.6.0              # Nested event loop support

# ----------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------
python-dateutil>=2.8.2           # Date parsing
structlog>=24.1.0                # Structured logging

# ================================================================
# INSTALLATION IN GITHUB ACTIONS
# ================================================================
# - name: Install Dependencies
#   run: |
#     python -m pip install --upgrade pip
#     pip install -r .pr-reviewer/requirements.txt
# ================================================================

# ================================================================
# NOTES
# ================================================================
# - Optimized for GitHub Actions runners (Ubuntu)
# - No development/testing-only packages
# - No Windows-specific packages
# - Fast installation (<2 minutes)
# ================================================================
